Refactor/#6/로그인 페이지 코드 구조 변경 및 InputField 생성
